karol/TripIdAsLong (#683)
* COnverted trip id from int to long

* trip id is now in format YYMMDDHHMMSS and rtc time is now calculated properly

+++ app/src/main/java/com/pitstop/bluetooth/dataPackages/TripInfoPackage.java
@@ -9,7 +9,7 @@ public class TripInfoPackage {
     }
 
     public String deviceId;
-    public int tripId; // device specific
+    public long tripId; // device specific
     public double mileage; // in kilometres
     public long rtcTime; // unix time in seconds
     public TripFlag flag;
